Make the extension mostly self-contained
