--- packages/lexical-playground/src/nodes/PageNode/index.tsx
@@ -1,12 +1,0 @@
-/**
- * Copyright (c) Meta Platforms, Inc. and affiliates.
- *
- * This source code is licensed under the MIT license found in the
- * LICENSE file in the root directory of this source tree.
- *
- */
-export * from './constants';
-export * from './PageContentNode';
-export * from './PageNode';
-export * from './PageSetupNode';
-export * from './types';
