+++ packages/lexical-playground/src/plugins/PagesExtension/pageSetup.ts
